+++ src/main/java/com/tools/codemos/dto/LeaderBoardDTO.java
@@ -7,11 +7,11 @@ import lombok.Setter;
 @Setter
 public class LeaderBoardDTO {
     private int id;
-    private double score;
+    private int score;
     private String loginId;
     private String nickname;
 
-    public LeaderBoardDTO(int id, double score, String loginId, String nickname) {
+    public LeaderBoardDTO(int id, int score, String loginId, String nickname) {
         this.id = id;
         this.score = score;
         this.loginId = loginId;
