+++ src/main/java/com/tools/codemos/dto/LeaderBoardRequest.java
@@ -1,21 +1,17 @@
 package com.tools.codemos.dto;
 
 public class LeaderBoardRequest {
-    private double score;
+    private int score;
     private String code;
-
-    public double getScore() {
+    public int getScore() {
         return score;
     }
-
     public void setScore(int score) {
         this.score = score;
     }
-
     public String getCode() {
         return code;
     }
-
     public void setCode(String code) {
         this.code = code;
     }
